refactor: improve code readability
- improve code readability
- add docs for some properties and methods
- split classes and methods to decrease the method/method number of lines
